+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -1,5 +1,6 @@
 # 4.1.0 (Unreleased)
 - Added `<ImageUpload />` component.
+- Style form components for WordPress 5.3.
 
 # 4.0.0